pyproject.toml
[tool.poetry]
name = "test-api"
version = "0.1.0"
description = ""
authors = ["Fedor Obolikhin <fedor.obolikhin@amonitoring.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
django-model-utils = "^4.3.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-environ = "^0.9.0"
drf-yasg = "^1.21.4"
djoser = "^2.1.0"
gunicorn = "^20.1.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
